lemma that has been said combining sentiment
