+++ smartteddy.basket.computer/web/smartteddydashboard/speechrecognition/models.py
@@ -2,8 +2,6 @@ from django.db import models
 from django.utils import timezone
 import datetime
 from datetime import date
-from django.core.exceptions import ValidationError
-from django.utils.translation import gettext_lazy as _
 
 POSITIVE = 'POS'
 NEGATIVE = 'NEG'
@@ -46,28 +44,18 @@ class SentimentLemma(models.Model):
         return f'{self.lemma_written_form} has a {self.sentiment_polarity} sentiment'
 
 
-def validate_week_number(value):
-    if value > 53:
-        raise ValidationError(_('%(value)s is higher than week number 52'), params={'value': value})
-    elif value < 1:
-        raise ValidationError(_('%(value)s is lower than week number 1'), params={'value': value})
-
-
-class LemmaCounter(models.Model):
-    id = models.PositiveBigIntegerField(primary_key=True)
-    # TODO create link (foreign key) for filtering a group of sentiment lemma's and recognized lemma's.
+class SaidLemma(models.Model):
+    id = models.BigAutoField(primary_key=True)
     word_lemma = models.CharField(max_length=100, default='<ukn>')
     date_said = models.DateTimeField('Date said on')
     week_number_said = models.PositiveSmallIntegerField(
         default=int(date.today().isocalendar()[1]),
-        validators=[validate_week_number]
     )
-
-    sentiment_lemma = models.ForeignKey(
+    sentiment = models.ForeignKey(
         SentimentLemma,
-        on_delete=models.CASCADE,
-        null=True,
+        on_delete=models.DO_NOTHING,
         blank=True,
+        null=True,
     )
 
     def was_said_last_week(self):
@@ -78,5 +66,14 @@ class LemmaCounter(models.Model):
     was_said_last_week.admin_order_field = 'date_said'
     was_said_last_week.short_description = 'Said last week'
 
-    def __dir__(self):
-        return f'{self.sentiment_lemma} is said on {self.date_said}'
+    def save(self, *args, **kwargs):
+        # TODO find a less heuristic for sentiment
+        search_sentiment = SentimentLemma.objects.filter(lemma_written_form=self.word_lemma)
+        if search_sentiment.count() == 1:
+            print(search_sentiment[0])
+            self.sentiment = search_sentiment[0]
+        self.week_number_said = int(self.date_said.isocalendar()[1])
+        super().save(*args, **kwargs)
+
+    def __str__(self):
+        return f"{self.word_lemma} is said on {self.date_said.strftime('%Y-%m-%d %H:%M')}"
