+++ smartteddy.basket.computer/web/smartteddydashboard/speechrecognition/views.py
@@ -2,7 +2,7 @@ from django.utils import timezone
 from rest_framework.response import Response
 from rest_framework.decorators import api_view, parser_classes
 from rest_framework.parsers import JSONParser
-from .models import LemmaCounter
+from .models import SaidLemma
 
 # TODO set minimum confidence level with feature toggle
 minimum_confidence = 0.0
@@ -16,7 +16,7 @@ def sentence(request):
         print(vosk_final_result)
         for partial_results in vosk_final_result['result']:
             if minimum_confidence <= partial_results['conf']:
-                lem = LemmaCounter(
+                lem = SaidLemma(
                     word_lemma=partial_results['word'],
                     date_said=timezone.now(),
                 )
